feat(xinfa): link compiler into trunk
