+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -21,7 +21,7 @@
       "source": {
         "kind": "file",
         "path": "framework/core/src/python/kungfu/agent/index.json",
-        "sha256": "2776f3c6ae9c2d4d43c84dab10a4c1fbd24e86b0db480adc924cce2b3558a38b"
+        "sha256": "627dcc93ee31dfa1ac679a827f10ab2aadb2d329fb778e1052f67cb5581e9182"
       },
       "evidence": [
         {
@@ -65,7 +65,7 @@
           "pointer": {
             "kind": "file",
             "path": "scripts/verify-agent-pack.mjs",
-            "sha256": "0134eedf22cee79198c9f9032f49b9aa85d9f04f88e620b3be0afd2549dac8f7"
+            "sha256": "63f4758ca0b4b656e24835cfcf01015aa35696565a5627e734846edc2e25053f"
           },
           "description": "Read-only verifier for installed agent pack completeness and command/doc consistency."
         },
