+++ .kungfu/project-cuts/sha256/20/20a17c24db92c8edaaa1e6bb09fb53ec5b8ba49b9287e9e57a137dc8c40e19c2/manifest.json
@@ -1,0 +1,1 @@
+{"atlas":{"compilerRoot":"sha256:d3e11301a51651127f06601cf2dffee73822fce50c7b767a71867c18c0d141b9","root":"sha256:60bebb0c6638b71551a21752651c4c3d51eb54d0c56da9e33c5003f59d5762c4","schema":"xinfa.atlas-ref/v1"},"compatibility":{"authorityMode":"native","existingRootsPreserved":true,"relations":[]},"conflicts":[],"cutRoot":"sha256:20a17c24db92c8edaaa1e6bb09fb53ec5b8ba49b9287e9e57a137dc8c40e19c2","episodeDelta":{"empty":true,"equivalenceProfileRoot":null,"nativeRoots":[],"schema":"kungfu.episode-delta-ref/v1","semanticRoot":null},"interpretation":{"policyRoots":["sha256:ff4318baf71b4ad79e609066626b0541e46af4faeeaa9f91260dd15322d7e2d8"],"protocolRoot":"sha256:d47894a6d80d6f406488bc0731440773cb902385e5711546ff9855aec27482b9","providerRoots":[],"schemaRoot":"sha256:3c4193cab90f3b1b48f504033b4d842b721d48c10259bd4ece8ea62a2fd9a055"},"omissions":[],"parentCutRoots":[],"project":{"id":"kungfu/xinfa-trunk-component","identityRoot":"sha256:f65213c9e6668dd15cac8a263ef933d96871670da708963dc49b07c8aebfffd9"},"schema":"project.cut/v1","sourceProjection":{"policyRoot":"sha256:ff4318baf71b4ad79e609066626b0541e46af4faeeaa9f91260dd15322d7e2d8","root":"sha256:ae1e364c106a891ef2ae2f7fe38449b47309d580ea9871ee33e88c5a5766a591","schema":"project.source-projection-ref/v1"},"unknowns":[],"visibility":"public"}
