+++ .kungfu/project-cuts/sha256/20/20a17c24db92c8edaaa1e6bb09fb53ec5b8ba49b9287e9e57a137dc8c40e19c2/receipt.json
@@ -1,0 +1,1 @@
+{"artifactDigest":"sha256:3af027c0e0093ee24e84ee84f8ccb06f9b80cd912d04addf38ed97649342839a","cutRoot":"sha256:20a17c24db92c8edaaa1e6bb09fb53ec5b8ba49b9287e9e57a137dc8c40e19c2","diagnostics":[],"publication":null,"receiptRoot":"sha256:9c36ef5d53407f563d56926c142bfd92ac7296252e3d227d4b4c5a52d0f36eb9","rootAlgorithm":"sha256-project-cut-canonical-json-v1","schema":"project.cut.receipt/v1","schemaRoot":"sha256:3c4193cab90f3b1b48f504033b4d842b721d48c10259bd4ece8ea62a2fd9a055","serializationRoot":"sha256:c22007d26d65c0286904c843b3b2d9f26fd7e39f55bb1cfe97b9ef1c5caf0b04","verdict":"valid"}
