+++ crates/trunk/src/main.rs
@@ -16,6 +16,7 @@ mod pins;
 mod plans;
 mod status;
 mod variant;
+mod xinfa_command;
 
 use std::env;
 use std::process::exit;
@@ -45,6 +46,7 @@ usage:
   kungfu-trunk repair-plan [--source <id>] [--episode <id>]
                                                plan storage repair (never writes)
   kungfu-trunk storage-status [--source <id>]  summarize native storage state
+  kungfu-trunk xinfa <command> [<args>...]      run the linked Xinfa compiler
   kungfu-trunk --version | --help
 
 envs live under <KF_HOME>/envs; the default env is named 'default'.
@@ -62,6 +64,7 @@ enum NativeCommand {
     GcPlan,
     RepairPlan,
     StorageStatus,
+    Xinfa,
 }
 
 struct NativeCommandSpec {
@@ -131,6 +134,13 @@ const NATIVE_COMMANDS: &[NativeCommandSpec] = &[
         summary: "summarize native storage state without CPython",
         section: "system-maintenance",
         visibility: "advanced",
+    },
+    NativeCommandSpec {
+        command: NativeCommand::Xinfa,
+        name: "xinfa",
+        summary: "compile workspace context into a verified Xinfa Atlas",
+        section: "agent-context",
+        visibility: "public",
     },
 ];
 
@@ -217,6 +227,7 @@ fn main() {
         Some("gc-plan") => plans::run_gc(&args[1..]),
         Some("repair-plan") => plans::run_repair(&args[1..]),
         Some("storage-status") => status::run(&args[1..]),
+        Some("xinfa") => xinfa_command::run(&args[1..]),
         Some("--version" | "-V" | "version") => {
             println!("kungfu-trunk {}", env!("CARGO_PKG_VERSION"));
             Ok(())
@@ -265,6 +276,7 @@ fn run_native(command: NativeCommand, args: &[String]) -> Result<(), String> {
         NativeCommand::GcPlan => plans::run_gc(args),
         NativeCommand::RepairPlan => plans::run_repair(args),
         NativeCommand::StorageStatus => status::run(args),
+        NativeCommand::Xinfa => xinfa_command::run(args),
     }
 }
 
@@ -665,7 +677,8 @@ mod tests {
                 "prewarm",
                 "repair-plan",
                 "storage-status",
-                "verify"
+                "verify",
+                "xinfa",
             ]
         );
     }
