+++ docs/adr/README.md
@@ -195,6 +195,7 @@ implemented and qualified or explicitly waived for that release.
 | [0123](ADR-0123-action-geometry-domain-profile-separation.md) | accepted | Action Geometry owns cross-domain responsibility separation and invariants; Domain Profiles own adopter-specific fields, lifecycle, policy, and presentation without gaining Fact or Episode authority |
 | [0124](ADR-0124-unified-invariant-verification-and-release-passport.md) | accepted | Fact and Episode invariants use one authority-bound registry, runner, evidence/object-receipt/passport vocabulary, successor gate, and fail-closed release projection without duplicating domain semantics |
 | [0125](ADR-0125-fact-episode-ontology-and-action-geometry.md) | accepted | Fact-Episode Ontology owns admitted state and realized causal occurrence; Action Geometry owns the three Pursuit, Atlas, and Warrant action coordinates over it |
+| [0126](ADR-0126-xinfa-trunk-linked-rust-component.md) | accepted | Xinfa remains an independently governed Rust component linked into the Kungfu trunk |
 | [SHIFU-0001](SHIFU-ADR-0001-cache-profile-contract-and-ownership.md) | accepted | Cache profiles are Shifu-owned contracts; inventories project instances and Buildchain owns process |
 | [SHIFU-0002](SHIFU-ADR-0002-local-artifact-catalog-and-safe-promotion.md) | accepted | Shifu and Kungfu product artifacts share provenance-aware, Git-safe local promotion semantics |
 | [SHIFU-0003](SHIFU-ADR-0003-uv-effective-lock-cache-enforcement.md) | accepted | Strict uv cache execution uses a disposable effective lock while canonical locks stay public |
